[feat] #57 feed Paging 설정

+++ data/src/main/java/com/dpm/sixpack/data/source/remote/datasoruce/FeedDataSource.kt
@@ -13,7 +13,7 @@ class FeedDataSource @Inject constructor(
 ) {
     suspend fun getFeeds(
         currentDate: String?,
-        userId: Int?,
+        userId: Long?,
         page: Int,
         size: Int,
     ): BaseResponse<FeedPageDto> = feedService.getFeeds(
